Calculate restore-progress via blockNo instead of slotNo
Large portions of slots may be empty. On the current jörmungandr testnet
the slot-based progress starts at around 96% and goes to 100%.

This commit uses blockNo instead, which perfectly matches the amount of
work the wallet has to do. This way progress starts at 0% (and still
goes to 100% at stable speed).